# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mix-Key Water Monitoring API",
        "description": "API Standard Documentation for Mix-Key Flood Alert System (Mae Chan, Chiang Rai)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Local Development API Gateway"
        }
    ],
    "paths": {
        "/stations": {
            "get": {
                "tags": [
                    "Guest APIs"
                ],
                "summary": "ดึงรายชื่อสถานีทั้งหมด (รองรับ GPS)",
                "operationId": "37a43e2a59cc97060c8c67b160738f06",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "ละติจูดของผู้ใช้ (เช่น 19.98765)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "ลองจิจูดของผู้ใช้ (เช่น 99.87654)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "รัศมีการค้นหา หน่วยกิโลเมตร (เช่น 10)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ดึงข้อมูลสำเร็จ"
                    }
                }
            }
        },
        "/stations/{id}": {
            "get": {
                "tags": [
                    "Guest APIs"
                ],
                "summary": "ดึงข้อมูลรายละเอียดสถานี",
                "operationId": "da19240f7794d804eaf8f8a4a542e513",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "รหัสสถานี",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ดึงข้อมูลสำเร็จ"
                    },
                    "404": {
                        "description": "ไม่พบข้อมูลสถานี"
                    }
                }
            }
        },
        "/stations/{id}/water-level": {
            "get": {
                "tags": [
                    "Guest APIs"
                ],
                "summary": "ดึงข้อมูลประวัติระดับน้ำย้อนหลัง (มี Aggregation + Pagination)",
                "operationId": "f6df1287d54fb3381e1557cca28f5ace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "รหัสสถานี",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "range",
                        "in": "query",
                        "description": "ช่วงเวลาของข้อมูล (6h, 1m, 3m, 6m)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "6h"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "หน้าข้อมูล (Pagination)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "จำนวนจุดข้อมูลต่อหน้า",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ดึงข้อมูลประวัติสำเร็จ"
                    }
                }
            }
        },
...
        "/telemetry/ingest": {
            "post": {
                "tags": [
                    "IoT Telemetry Ingestion"
                ],
                "summary": "ช่องทางรับข้อมูลดิบจาก Node-RED/Sensor",
                "operationId": "cb3bb89a11c460312565716b8ace48ff",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "รหัสลับความปลอดภัยอุปกรณ์",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sensor_id",
                                    "water_level_cm",
                                    "battery_voltage"
                                ],
                                "properties": {
                                    "sensor_id": {
                                        "type": "string",
                                        "example": "PT-01"
                                    },
                                    "timestamp": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-06-08T16:15:00+07:00"
                                    },
                                    "water_level_cm": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 245
                                    },
                                    "battery_voltage": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3.82
                                    },
                                    "signal_strength_rssi": {
                                        "type": "integer",
                                        "example": -72
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "บันทึกข้อมูลเรียบร้อยแล้ว"
                    },
                    "401": {
                        "description": "API Key ไม่ถูกต้อง"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "เข้าสู่ระบบสตาฟฟ์/แอดมิน เพื่อรับ Bearer Token",
                "operationId": "3b578bde47faf5711ca9c22c8d3917ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mawin.staff@maejhan.go.th"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "securepassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ล็อกอินสำเร็จ"
                    },
                    "422": {
                        "description": "ข้อมูลตรวจสอบไม่ผ่าน"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "ออกจากระบบ (ทำลาย Token)",
                "operationId": "75430f4aa001cb1457460c229448295c",
                "responses": {
                    "200": {
                        "description": "ออกจากระบบสำเร็จ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/staff/stations/{id}/telemetry": {
            "get": {
                "tags": [
                    "Staff APIs"
                ],
                "summary": "ดึงข้อมูลสถานะการทำงานจริงของเครื่อง (สตาฟฟ์เท่านั้น)",
                "operationId": "76155da526279cd7421c854aa2a5dd1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "รหัสสถานี",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ดึงข้อมูลสำเร็จ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/staff/stations/{id}/thresholds": {
            "patch": {
                "tags": [
                    "Staff APIs"
                ],
                "summary": "ปรับเปลี่ยนเกณฑ์เตือนภัยระดับน้ำ (สตาฟฟ์เท่านั้น)",
                "operationId": "bd6797494ed4b2f833468050a6e6cc20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "รหัสสถานี",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "warning_datum_m": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.2
                                    },
                                    "critical_datum_m": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3.1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "อัปเดตระดับเตือนภัยสำเร็จ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/staff/stations/{id}/maintenance": {
            "get": {
                "tags": [
                    "Staff APIs"
                ],
                "summary": "เรียกดูประวัติการซ่อมบำรุงของสถานี (สตาฟฟ์เท่านั้น)",
                "operationId": "212dc3bc21211528948c83a742a1bd82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "รหัสสถานี",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ดึงประวัติสำเร็จ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter Bearer Token to access protected Staff/Admin endpoints",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}